장바구니 페이지, 구매자/판매자 주문 관련 페이지 완성, 통계 dto 설계

+++ cake-quake-back/src/main/java/com/cakequake/cakequakeback/order/dto/seller/SellerOrderDetail.java
@@ -54,7 +54,7 @@ public class SellerOrderDetail {
         private Integer unitPrice;
 
         /** CakeOrderItem 엔티티의 subTotalPrice 필드 (소계 금액) */
-        private BigDecimal subTotalPrice;
+        private Integer subTotalPrice;
 
         /** CakeItem 엔티티의 thumbnailImageUrl 필드 (썸네일 URL) */
         private String thumbnailImageUrl;
